requirements.txt
setuptools==40.5.0
PyYAML>=3.10, <4.0
click>=5.1, <5.2
click-log==0.1.1
requests
PyGithub
boto3>1.9, <1.10
